requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.3
botocore==1.38.3
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
jmespath==1.0.1
kombu==5.5.1
pillow==11.2.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
